resolve CI errors
